+++ src/Notifications/HasLocalNotification.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ikromjon\LocalNotifications\Notifications;
+
+interface HasLocalNotification
+{
+    public function toLocalNotification(object $notifiable): LocalNotificationMessage;
+}